chore(BaseForm): moving file directories
