+++ packages/hedgehog-lab/src/components/Base/Input/IBaseInputProps.ts
@@ -2,7 +2,7 @@ import {ControllerRenderProps} from "react-hook-form/dist/types/controller";
 import {BaseTextFieldProps} from "@mui/material";
 import React from "react";
 
-export interface IFormProps extends BaseTextFieldProps {
+export interface IBaseInputProps extends BaseTextFieldProps {
     name: string,
     field?: ControllerRenderProps,
     control?: any,
