+++ packages/hedgehog-lab/src/components/Form/Base/BaseAuthForm/BaseAuthForm.tsx
@@ -1,7 +1,7 @@
 import {useFormContext} from "react-hook-form";
-import EmailInput from "../../Base/Input/Email/EmailInput";
+import EmailInput from "../../../Base/Input/Email/EmailInput";
 import {Box} from "@mui/material";
-import PasswordInput from "../../Base/Input/Password/PasswordInput";
+import PasswordInput from "../../../Base/Input/Password/PasswordInput";
 import * as React from "react";
 
 const BaseAuthForm: React.FC = (prop): React.ReactElement => {
