+++ packages/hedgehog-lab/src/components/User/Auth/AuthForget/AuthForget.tsx
@@ -1,13 +1,13 @@
 import * as React from "react";
-import BaseForm from "../../../Form/Base/BaseForm";
-import BaseAction from "../../../Form/Base/BaseAction";
+import BaseForm from "../../../Form/Base/BaseForm/BaseForm";
+import BaseAuthAction from "../../../Form/Base/BaseAuthAction/BaseAuthAction";
 import {Box} from "@mui/material";
 import {useFormContext} from "react-hook-form";
 import EmailInput from "../../../Base/Input/Email/EmailInput";
 
 const ForgetAction = () =>
     (
-        <BaseAction
+        <BaseAuthAction
             action={{text: 'Forget'}}
             signOrLogin={{text: 'Already a member?', actionText: 'Log in', action: 'login'}}
             forget={{text: 'Forget password?'}}
