+++ packages/hedgehog-lab/src/components/User/Auth/AuthSign/AuthSign.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
-import BaseForm from "../../../Form/Base/BaseForm";
-import BaseAction from "../../../Form/Base/BaseAction";
-import BaseAuthForm from "../../../Form/Base/BaseAuthForm";
+import BaseForm from "../../../Form/Base/BaseForm/BaseForm";
+import BaseAuthAction from "../../../Form/Base/BaseAuthAction/BaseAuthAction";
+import BaseAuthForm from "../../../Form/Base/BaseAuthForm/BaseAuthForm";
 import {Box} from "@mui/material";
 import {useFormContext} from "react-hook-form";
 import UserNameInput from "../../../Base/Input/UserName/UserNameInput";
 
 const SignAction = () =>
     (
-        <BaseAction
+        <BaseAuthAction
             action={{text: 'Sign In'}}
             signOrLogin={{text: 'Already a member?', actionText: 'Log in', action: 'login'}}
         />
